Add ReplaceCommand
